requirements.txt
absl-py==0.9.0
appdirs==1.4.4
astor==0.8.1
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cryptography==3.1
cssselect==1.1.0
cycler==0.10.0
future==0.18.2
gast==0.4.0
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.7.0
joblib==0.16.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lxml==4.5.2
Markdown==3.2.2
matplotlib==3.3.1
nltk==3.5
numpy==1.19.1
opencv-python==4.4.0.42
pandas==1.1.1
pdfminer.six==20200726
Pillow==7.2.0
protobuf==3.13.0
pycparser==2.20
pymongo==3.11.0
pyparsing==2.4.7
pytesseract==0.3.5
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
requests==2.24.0
six==1.15.0
sortedcontainers==2.2.2
soupsieve==2.0.1
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.48.2
urllib3==1.25.10
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0

